fix some funcs to return bool

+++ database/src/PasteDBManager.cpp
@@ -35,6 +35,8 @@ std::shared_ptr<dataFormat> PasteDBManager::getPaste(const std::string& hash) {
 
 bool PasteDBManager::deletePaste(const std::string& hash) {
     conditionMapFormat conditionMap = createHashConditionMap(hash);
+    if (getPaste(hash) == nullptr)
+        return false;
     return deleteByPK(conditionMap, PASTE_TABLE_NAME);
 }
 
@@ -42,6 +44,14 @@ bool PasteDBManager::updatePaste(const std::string& hash,
                                  const dataFormat& newParamsMap) {
     conditionMapFormat conditionMap = createHashConditionMap(hash);
     return updateByPK(conditionMap, newParamsMap, PASTE_TABLE_NAME);
+}
+
+bool PasteDBManager::updateText(const std::string& hash,
+                                const std::string& newText) {
+    if (!checkHash(hash))
+        return false;
+    conditionMapFormat conditionMap = createHashConditionMap(hash);
+    return updateByPK(conditionMap, {{"text", newText}}, PASTE_TABLE_NAME);
 }
 
 bool PasteDBManager::checkHash(const std::string& hash) {
